+++ CHANGELOG.md
@@ -13,10 +13,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (https://github.com/mad-lab-fau/tpcp/pull/95)
 - The `PyTestSnapshotTest` class does now support comparing dataframes with datetime columns.
   (https://github.com/mad-lab-fau/tpcp/pull/97)
+- New example about caching.
+  (https://github.com/mad-lab-fau/tpcp/pull/98)
 
 ### Changed
 
 - In line with numpy and some other packages, we drop Python 3.8 support
+
 
 ## [0.24.0] - 2023-09-08